requirements.txt
diff-match-patch==20121119
dj-database-url==0.4.1
Django==1.10.1
django-extensions==1.7.4
django-import-export==0.5.0
gunicorn==19.6.0
six==1.10.0
tablib==0.11.2
whitenoise==3.2.1
